% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_path.R
\name{build_path}
\alias{build_path}
\title{build_path}
\usage{
build_path(FilePath, Silent = TRUE)
}
\arguments{
\item{FilePath}{A character string with the target folder path}

\item{Silent}{True by default, if set to FALSE it shows the address of the folder just created.}
}
\value{
Path address just built.
}
\description{
build_path builds the entire folder FilePath provided. If the FilePath does not exist, it builds it without error. It is effectively a user-friendly wrapper to the base function dir.create.
}
\examples{
build_path("C:/Temp/data")
}

