% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTempFolder.R
\name{CleanTempFolder}
\alias{CleanTempFolder}
\title{CleanTempFolder}
\usage{
CleanTempFolder()
}
\value{
Cleaned R Temp Folder.
}
\description{
This function will clean all the csv files created using ViewXL in the temp folder. This might take a while to clean if not done regularly.
}
\examples{
CleanTempFolder()
}

