% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PromptAsTime.R
\name{PromptAsTime}
\alias{PromptAsTime}
\title{PromptAsTime}
\usage{
PromptAsTime(On)
}
\arguments{
\item{On}{set On to TRUE (Add time to prompter) or FALSE (use default prompter).}
}
\value{
The Prompter in Rstudio will now include the time.
}
\description{
This changes Rstudio's prompt at the bottom to reflect time. Particularly useful for timing functions when executing long scripts.
}
\examples{
\dontrun{
PromptAsTime(TRUE)
x <- 100
Sys.sleep(3)
#' x*x
print(x)
}
}
