% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTempFolder.R
\name{CleanTempFolder}
\alias{CleanTempFolder}
\title{CleanTempFolder}
\usage{
CleanTempFolder(RTempFileLocation)
}
\arguments{
\item{User}{Name. If left blank it will use the location defined by tempdir() in R. Else, manually define where temporary R files are stored.}
}
\value{
CLeaned R Temp Folder.
}
\description{
This function will clean all the Rtmp files created. This might take a while to clean if not done regularly. This can be used to clean the R temporary folder specifically when using ViewXL.
}
\examples{
CleanTempFolder()
}

