% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PromptAsTime.R
\name{PromptAsTime}
\alias{PromptAsTime}
\title{PromptAsTime}
\usage{
PromptAsTime(On)
}
\arguments{
\item{Set}{On to TRUE (Add time to prompter) or FALSE (use default prompter).}
}
\value{
The Prompter in Rstudio will now include the time.
}
\description{
This changes Rstudio's prompt at the bottom to reflect time. Useful for timing functions with ease.
}
\examples{
PromptAsTime(TRUE)
}

